move all use case classes into the usecase package.
