+++ src/main/kotlin/dependencyinjection/projectModule.kt
@@ -4,11 +4,7 @@ import org.example.data.CsvMealParser
 import org.example.data.CsvMealReader
 import org.example.data.CsvRepositoryImpl
 import org.example.logic.CsvRepository
-import org.example.logic.EasyFoodSuggestionUseCase
-import org.example.logic.KetoDietMealHelperUseCase
-import org.example.logic.SweetMealWithoutEggUseCase
 import org.example.utils.MealSearchIndex
-import org.example.presentation.FoodApplicationUI
 import org.example.presentation.io.ConsoleIO
 import org.example.presentation.io.RealConsoleIO
 import org.koin.dsl.module
