+++ src/main/kotlin/logic/usecase/ExploreCountriesFoodCultureUseCase.kt
@@ -1,5 +1,6 @@
-package org.example.logic
+package org.example.logic.usecase
 
+import org.example.logic.CsvRepository
 import org.example.model.Meal
 import org.example.utils.bitapFuzzySearch
 import org.example.utils.preprocessForFuzzySearch
