+++ src/main/kotlin/logic/usecase/GetMealsByAddDateUseCase.kt
@@ -1,6 +1,7 @@
-package org.example.logic
+package org.example.logic.usecase
 
 import kotlinx.datetime.LocalDate
+import org.example.logic.CsvRepository
 import org.example.model.Meal
 import org.example.utils.MealAppException
