+++ src/main/kotlin/logic/usecase/PlayIngredientGameUseCase.kt
@@ -1,4 +1,4 @@
-package org.example.logic
+package org.example.logic.usecase
 
 import org.example.model.IngredientGameResult
 import org.example.model.Meal
