+++ src/main/kotlin/logic/usecase/SweetMealWithoutEggUseCase.kt
@@ -1,5 +1,6 @@
-package org.example.logic
+package org.example.logic.usecase
 
+import org.example.logic.ReactionProvider
 import org.example.model.Meal
 import org.example.utils.MealAppException
