+++ src/main/kotlin/presentation/controller/EasyFoodSuggestionController.kt
@@ -1,6 +1,6 @@
 package org.example.presentation.controller
 
-import org.example.logic.EasyFoodSuggestionUseCase
+import org.example.logic.usecase.EasyFoodSuggestionUseCase
 import org.example.presentation.io.ConsoleIO
 
 class EasyFoodSuggestionController(
